Move friendship functions from UserService to FriendshipService
