+++ UniVerseAPI/universe/src/main/java/com/unidev/universe/services/UserService.java
@@ -2,7 +2,7 @@ package com.unidev.universe.services;
 
 import com.unidev.universe.entities.User;
 import com.unidev.universe.repository.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
@@ -13,8 +13,8 @@ import java.util.List;
 import java.util.Optional;
 
 @Service
+@RequiredArgsConstructor
 public class UserService implements UserDetailsService {
-    @Autowired
     private UserRepository userRepository;
 
     @Override
@@ -34,79 +34,5 @@ public class UserService implements UserDetailsService {
 
     public List<User> getAllUsers() {
         return userRepository.findAll();
-    }
-
-    public List<User> getOnlineFriends(String username) {
-        Optional<User> optionalUser = Optional.ofNullable(userRepository.findByUsername(username));
-        if (optionalUser.isPresent()) {
-            List<User> onlineFriends = new ArrayList<>();
-
-            List<String> friendUsernames = getFriendUsernames(username);
-
-            for (String friendUsername : friendUsernames) {
-                Optional<User> friendEntity = Optional.ofNullable(userRepository.findByUsername(friendUsername));
-                if (friendEntity.isPresent()) {
-                    User friend = friendEntity.get();
-                    if (friend.getEmail() != null) {
-                        onlineFriends.add(friend);
-                    }
-                }
-            }
-            return onlineFriends;
-        }
-        return new ArrayList<>();
-    }
-
-    public List<String> getFriendUsernames(String username) {
-        Optional<User> optionalUser = Optional.ofNullable(userRepository.findByUsername(username));
-        List<String> friendUsernames = new ArrayList<>();
-
-        if (optionalUser.isPresent()) {
-            User user = optionalUser.get();
-            List<User> allUsers = userRepository.findAll();
-
-            for (User friend : allUsers) {
-                // Checking if the friend's email and username are not null
-                if (!friend.getEmail().isEmpty() && !friend.getUsername().equals(user.getUsername())) {
-                    friendUsernames.add(friend.getUsername());
-                }
-            }
-        }
-
-        return friendUsernames;
-    }
-
-    public boolean areFriends(String usernameUser1, String usernameUser2) {
-        Optional<User> optionalUser1 = Optional.ofNullable(userRepository.findByUsername(usernameUser1));
-        Optional<User> optionalUser2 = Optional.ofNullable(userRepository.findByUsername(usernameUser2));
-
-        if (optionalUser1.isPresent() && optionalUser2.isPresent()) {
-            boolean user1IsFriendOfUser2 = userRepository.checkFriendship(usernameUser1, usernameUser2);
-            boolean user2IsFriendOfUser1 = userRepository.checkFriendship(usernameUser2, usernameUser1);
-
-            return user1IsFriendOfUser2 && user2IsFriendOfUser1;
-        }
-
-        return false;
-    }
-
-    public void addFriend(String loggedInUser, String profileUser) {
-        Optional<User> optionalLoggedInUser = Optional.ofNullable(userRepository.findByUsername(loggedInUser));
-        Optional<User> optionalProfileUser = Optional.ofNullable(userRepository.findByUsername(profileUser));
-
-        if (optionalLoggedInUser.isPresent() && optionalProfileUser.isPresent()) {
-            userRepository.addFriendByUsername(loggedInUser, profileUser);
-            userRepository.addFriendByUsername(profileUser, loggedInUser);
-        }
-    }
-
-    public void removeFriend(String loggedInUser, String profileUser) {
-        Optional<User> optionalLoggedInUser = Optional.ofNullable(userRepository.findByUsername(loggedInUser));
-        Optional<User> optionalProfileUser = Optional.ofNullable(userRepository.findByUsername(profileUser));
-
-        if (optionalLoggedInUser.isPresent() && optionalProfileUser.isPresent()) {
-            userRepository.removeFriendByUsername(loggedInUser, profileUser);
-            userRepository.removeFriendByUsername(profileUser, loggedInUser);
-        }
     }
 }
